--- __init__.py
@@ -1,11 +1,0 @@
-__title__ = 'Lavalink'
-__author__ = 'Luke & William'
-__license__ = 'MIT'
-__copyright__ = 'Copyright 2018 Luke & William'
-__version__ = '2.0.2.3'
-
-from .audio_track import *
-from .client import *
-from .players import *
-from .utils import *
-from .web_socket import *
